#176: address inner-loop review — DeviceType case, reuse FileClient HTTP, SocketException, narrating comments
- DeviceType.Mobile/Desktop call sites aligned with PascalCase enum entries
- RendezvousClient deleted; sendHello moved onto FileClient to reuse its HttpClient (discovery.md §Cross-layer concerns — no new HTTP stack)
- MdnsDiscoveryJmdns.bindAddresses() wraps NetworkInterface.getNetworkInterfaces() in try/catch SocketException (Windows portability)
- two narrating comments removed
- NsdTxtPropsTest, apple FileServer*Test, MdnsDiscoveryTest aligned with new txtProps(fp) and configuredPort signatures
