+++ composeApp/src/appleTest/kotlin/com/tubetoast/tether/network/FileServerConcurrencyTest.kt
@@ -70,7 +70,7 @@ class FileServerConcurrencyTest {
         val configDir = newTempDir()
         val temp = TempDataStore().also { cleanupTempStores += it }
         return FileServer(
-            port = 0,
+            configuredPort = 0,
             downloadsDir = downloadsDir,
             trustedDeviceStore = DefaultTrustedDeviceStore(temp.dataStore),
             deviceKeyPair = DeviceKeyPair(configDir),
